working spawning islands
